Rearranging tests + adding more cases
