+++ engine/brain/src/test/mod.rs
@@ -1,0 +1,3 @@
+mod move_comparison;
+mod best_move;
+mod clearcut;
